Use relative paths and move container component in separate file
