+++ lib/main.dart
@@ -1,8 +1,8 @@
 import 'package:flutter/material.dart';
-import 'package:gymbuddies/pages/home.dart';
-import 'package:gymbuddies/pages/training.dart';
-import 'package:gymbuddies/pages/calender.dart';
-import 'package:gymbuddies/pages/contacts.dart';
+import 'pages/home.dart';
+import 'pages/training.dart';
+import 'pages/calender.dart';
+import 'pages/contacts.dart';
 
 void main() {
   runApp(const MyApp());
