+++ lib/pages/home.dart
@@ -1,7 +1,8 @@
 import 'package:flutter/material.dart';
-import 'package:gymbuddies/colors.dart';
-import 'package:gymbuddies/common_layout.dart';
-import 'package:gymbuddies/pages/training.dart';
+import '../colors.dart';
+import '../common_layout.dart';
+import '../components/container_with_text.dart';
+import '../pages/training.dart';
 
 class HomePage extends StatelessWidget {
   const HomePage({super.key});
@@ -11,71 +12,6 @@ class HomePage extends StatelessWidget {
     return const CommonLayout(
       body: HomePageBody(),
       selectedIndex: 0,
-    );
-  }
-}
-
-class ContainerWithHeading extends StatelessWidget {
-  final String heading;
-  final List<String> containerText;
-  final bool darkMode;
-
-  const ContainerWithHeading({
-    required this.heading,
-    required this.containerText,
-    required this.darkMode,
-    super.key,
-  });
-
-  @override
-  Widget build(BuildContext context) {
-    return Column(
-      crossAxisAlignment: CrossAxisAlignment.start,
-      children: [
-        Row(
-          children: [
-            Text(
-              heading,
-              style: TextStyle(
-                  fontSize: 24.0,
-                  fontWeight: FontWeight.bold,
-                  color: AppColors.getSecondaryColor(darkMode),
-                  decoration: TextDecoration.underline,
-                  decorationColor: AppColors.getAccentColor(darkMode)),
-            ),
-          ],
-        ),
-        const Padding(padding: EdgeInsetsDirectional.symmetric(vertical: 5.0)),
-        Row(
-          children: [
-            Container(
-              padding:
-                  const EdgeInsets.symmetric(horizontal: 20.0, vertical: 10.0),
-              height: MediaQuery.of(context).size.width * 0.45,
-              width: MediaQuery.of(context).size.width * 0.87,
-              decoration: BoxDecoration(
-                color: AppColors.getSecondaryColor(darkMode),
-                boxShadow: [
-                  BoxShadow(
-                      color: AppColors.getShadowColor(darkMode),
-                      spreadRadius: 3,
-                      blurRadius: 3),
-                ],
-              ),
-              child: Column(
-                mainAxisAlignment: MainAxisAlignment.spaceAround,
-                crossAxisAlignment: CrossAxisAlignment.start,
-                children: containerText.map((String text) {
-                  return Text(
-                    text,
-                    style: TextStyle(color: AppColors.getTextColor(darkMode)),
-                  );
-                }).toList(),
-              ),
-            ),
-          ],
-        ),
-      ],
     );
   }
 }
@@ -108,7 +44,7 @@ class HomePageBody extends StatelessWidget {
         Row(
           mainAxisAlignment: MainAxisAlignment.center,
           children: [
-            ContainerWithHeading(
+            ContainerWithText(
               heading: 'Upcoming',
               containerText: upcomingTrainingSessions,
               darkMode: isDarkMode,
@@ -118,7 +54,7 @@ class HomePageBody extends StatelessWidget {
         Row(
           mainAxisAlignment: MainAxisAlignment.center,
           children: [
-            ContainerWithHeading(
+            ContainerWithText(
               heading: 'Last Trainings',
               containerText: lastTrainingSessions,
               darkMode: isDarkMode,
